+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Change `psakefile.ps1` to detect Azure Pipelines correctly.
 - Added PowerShell Gallery badge to `README.md`.
+- Update URL Encoding to use more modern URL encoding method
+  `[System.Web.Security.AntiXss.AntiXssEncoder]::UrlEncode()` - fixes
+  [Issue #7](https://github.com/PlagueHO/PSAuth/issues/7).
 
 ## 0.1.2.36